Extend mypy plugin with decorator base class injection and fix retry_delay_seconds type
Enhance the Protean mypy plugin to understand decorator-based class
registration (@domain.aggregate, @domain.entity, etc.) by injecting
base class symbols during type analysis. This eliminates false mypy
errors for methods like raise_(), to_dict(), and auto-injected
attributes like `id` on decorated classes.

Key technical decisions:
- Use get_customize_class_mro_hook instead of get_class_decorator_hook
  because @dataclass_transform() on Domain methods causes mypy to
  bypass the decorator hook entirely
- Copy base class symbols into the class namespace rather than modifying
  info.bases to avoid metaclass conflicts with pydantic's ModelMetaclass
- Match decorators by AST inspection (MemberExpr/CallExpr) since name
  resolution may be incomplete during MRO calculation

Plugin additions:
- DECORATOR_BASE_CLASS_MAP covering all 15 domain element decorators
- _extract_decorator_name_from_expr for AST-based decorator matching
- _inject_base_class with symbol copying from base class hierarchy
- _maybe_inject_auto_id for aggregate/entity id field injection
- PROTEAN_MYPY_DEBUG=1 environment variable for diagnostic output

Testing (53 tests):
- 6 new mypy fixture files for decorator type verification
- 12 new integration tests running mypy on decorator fixtures
- 20 new unit tests for decorator mapping, extraction, and hooks

Infrastructure:
- VS Code settings: disable Pylance type checking, configure mypy extension
- CI: add dedicated mypy plugin test step before full test suite
- Makefile: add `typecheck` target
- Documentation: type-checking guide at docs/guides/type-checking.md

+++ src/protean/server/subscription/stream_subscription.py
@@ -37,7 +37,7 @@ class StreamSubscription(BaseSubscription):
         messages_per_tick: Optional[int] = None,
         blocking_timeout_ms: Optional[int] = None,
         max_retries: Optional[int] = None,
-        retry_delay_seconds: Optional[int] = None,
+        retry_delay_seconds: Optional[float] = None,
         enable_dlq: Optional[bool] = None,
     ) -> None:
         """
@@ -53,7 +53,7 @@ class StreamSubscription(BaseSubscription):
                 Defaults to config value or 5000.
             max_retries (int, optional): Maximum number of retries before moving to DLQ.
                 Defaults to config value or 3.
-            retry_delay_seconds (int, optional): Delay between retries in seconds.
+            retry_delay_seconds (float, optional): Delay between retries in seconds.
                 Defaults to config value or 1.
             enable_dlq (bool, optional): Whether to use a dead letter queue.
                 Defaults to config value or True.
@@ -78,10 +78,10 @@ class StreamSubscription(BaseSubscription):
             if max_retries is not None
             else int(stream_config.get("max_retries", 3))
         )
-        resolved_retry_delay_seconds: int = (
+        resolved_retry_delay_seconds: float = (
             retry_delay_seconds
             if retry_delay_seconds is not None
-            else int(stream_config.get("retry_delay_seconds", 1))
+            else float(stream_config.get("retry_delay_seconds", 1))
         )
         resolved_enable_dlq: bool = (
             enable_dlq
@@ -104,7 +104,7 @@ class StreamSubscription(BaseSubscription):
         self.stream_category = stream_category
         self.blocking_timeout_ms: int = resolved_blocking_timeout_ms
         self.max_retries: int = resolved_max_retries
-        self.retry_delay_seconds: int = resolved_retry_delay_seconds
+        self.retry_delay_seconds: float = resolved_retry_delay_seconds
         self.enable_dlq: bool = resolved_enable_dlq
 
         # Consumer name for Redis Streams (unique per consumer instance)
